create file metrics.rs
