+++ linera-views/src/test_utils/mod.rs
@@ -2,6 +2,7 @@
 // SPDX-License-Identifier: Apache-2.0
 
 pub mod test_views;
+
 use std::{
     collections::{BTreeMap, BTreeSet, HashSet},
     fmt::Debug,
